Initiat Task
A Registration Page with fields like - Name, Email, Password, and Picture so that users can register on your website.
Allow the registered user to login to your website and the authenticated user will be redirected to a dashboard page that says “Welcome” along with the “Name” of the user and the profile picture of the user.
The dashboard page will also have a menu with- “dashboard, profile, and search”
Clicking on the profile picture will take the user to a page where the user can update his data like - Name, Email, and Profile Picture. Users can also update their passwords on the same page.
Clicking on Search will take the user to a search page where the user can search for any image/video using Pixabay API (for details on the Pixabay API key, follow the documentation). A successful search will return the Image/Video from Pixabay and show them on the search page.

Please note to do it in CodeIgniter.